--- HW3.1/HW3.1/ViewController.swift
@@ -1,37 +1,0 @@
-//
-//  ViewController.swift
-//  HW3.1
-//
-//  Created by Peter K on 05.03.2021.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-    
-    
-    @IBAction func additionButton(_ sender: Any) {
-        
-        self.alert(title: "Enter two numbers", message: "", style: .alert)
-        
-        }
-    func alert(title: String, message: String, style: UIAlertController.Style) {
-        
-        let alert = UIAlertController(title: title, message: message, preferredStyle: style)
-        let action = UIAlertAction(title: "Ok", style: .default, handler: nil)
-        
-        alert.addTextField(configurationHandler: nil)
-        alert.addTextField(configurationHandler: nil)
-        alert.addAction(action)
-        
-        self.present(alert, animated: true, completion: nil)
-    }
-    
-
-}
-
